% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_na.R
\name{generate_na}
\alias{generate_na}
\title{Adds Missing Values to a Data Set}
\usage{
generate_na(x, p = 0.1, seed = NULL)
}
\arguments{
\item{x}{\code{data.frame} or \code{tibble}.}

\item{p}{Proportion of missing values to approximately add to each column of
\code{x}. Default value is 0.1.}

\item{seed}{An integer seed.}
}
\value{
\code{x} with missing values.
}
\description{
Takes a \code{data.frame} or a \code{tibble} and replaces
randomly part of the values by missing values.
}
\examples{

head(generate_na(iris))

}
