% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_factor.R
\name{impute_factor}
\alias{impute_factor}
\title{Majority Imputation}
\usage{
impute_factor(x)
}
\arguments{
\item{x}{A factor vector possibly containing missing values.}
}
\value{
A factor vector of the same length as \code{x} but without missing
values.
}
\description{
Fills missing values of a factor vector by using the majority
rule. If there are several levels with the same number of votes, samples one
at random  and fills the missing values
}
\examples{

x <- as.factor(c(1,1,2,2,3,3,3))
x[3] <- NA
impute_factor(x)

}
