% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_median.R
\name{impute_median}
\alias{impute_median}
\title{Median Imputation}
\usage{
impute_median(x)
}
\arguments{
\item{x}{A numeric vector possibly containing missing values.}
}
\value{
A numeric vector of the same length as \code{x} but without missing
values.
}
\description{
Fills missing values of a numeric vector by using its median
value.
}
\examples{

impute_median(c(NA, 0, 1, 0, 1))

}
