% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_tune.R
\name{miss_tune}
\alias{miss_tune}
\title{Imputation of Missing Values by Automatic Tuned Chained Tree Ensembles}
\usage{
miss_tune(x_miss, max_iter = 10L, seed = NULL, num_trees = 200,
  tune_length = 12, verbose = TRUE)
}
\arguments{
\item{x_miss}{A \code{data.frame} or \code{tibble} with missing values to
impute.}

\item{max_iter}{Maximum number of chaining iterations.}

\item{seed}{Integer seed to initialize the random generator.}

\item{num_trees}{Number of trees passed to \code{train} function of the
\code{caret} package.}

\item{tune_length}{the number of random sets of hyper-parameters passed to
the \code{trainControl} function of the \code{caret} package.}

\item{verbose}{Boolean. FALSE (default) to print nothing, TRUE to print
the OOB prediction error per iteration and variable (1 minus R-squared for
regression).}
}
\value{
A class with the imputed data having the smaller OOB error, and all
the OOB errors from the iterations of the algorithm.
}
\description{
Uses the \code{caret} and the \code{ranger} packages to do fast
missing value imputation by automatic tuned chained tree ensembles, see
[1, 2]. The resampling method used for the selection of the automatic
tuning is five folds cross validation.
The iterative chaining stops as soon as \code{max_iter} is reached or if the
average out-of-bag estimate of performance stops improving.
In the latter case, except for the first iteration, the second last
(i.e. best) imputed data is returned.
}
\examples{

\dontrun{
iris_na <- generate_na(iris)
iris_imp <- miss_tune(iris_na)
head(iris_imp)
head(iris_na)
}
}
\references{
[1] Wright, M. N. & Ziegler, A. (2016). ranger: A Fast Implementation of
Random Forests for High Dimensional Data in C++ and R. Journal of Statistical
Software, in press. http://arxiv.org/abs/1508.04409.

[2] Stekhoven, D.J. and Buehlmann, P. (2012). 'MissForest - nonparametric
missing value imputation for mixed-type data', Bioinformatics, 28(1) 2012,
112-118, doi: 10.1093/bioinformatics/btr597
}
