% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missTune-package.R
\docType{package}
\name{missTune}
\alias{missTune}
\alias{_PACKAGE}
\alias{missTune-package}
\title{missTune: Fast Imputation of Missing Values by Automatic Tuned Chained Tree Ensembles.}
\description{
Alternative implementation of the 'missRanger' package using
tuned Random Forests. missRanger is also an alternative implementation of the
'MissForest' algorithm used to impute mixed-type data sets by chaining tree
ensembles, introduced by Stekhoven, D.J. and Buehlmann, P. (2012)
<doi:10.1093/bioinformatics/btr597>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kvantas/missTune}
  \item Report bugs at \url{https://github.com/kvantas/missTune/issues}
}

}
\author{
\strong{Maintainer}: Konstantinos Vantas \email{kon.vantas@gmail.com}

}
\keyword{internal}
